+++ packages/lumx-icons/override/alias-icons.js
@@ -1,0 +1,486 @@
+module.exports = {
+    'account-cowboy-hat': [
+        // retro-compat v4 aliases:
+        'cowboy',
+    ],
+    'account-hard-hat': [
+        // retro-compat v4 aliases:
+        'worker',
+    ],
+    'account-music': [
+        // retro-compat v4 aliases:
+        'artist',
+    ],
+    'account-music-outline': [
+        // retro-compat v4 aliases:
+        'artist-outline',
+    ],
+    'account-voice': [
+        // retro-compat v4 aliases:
+        'voice',
+    ],
+    'apple-airplay': [
+        // retro-compat v4 aliases:
+        'airplay',
+    ],
+    'badge-account': [
+        // retro-compat v4 aliases:
+        'account-badge',
+    ],
+    'badge-account-alert': [
+        // retro-compat v4 aliases:
+        'account-badge-alert',
+    ],
+    'badge-account-alert-outline': [
+        // retro-compat v4 aliases:
+        'account-badge-alert-outline',
+    ],
+    'badge-account-horizontal': [
+        // retro-compat v4 aliases:
+        'account-badge-horizontal',
+    ],
+    'badge-account-horizontal-outline': [
+        // retro-compat v4 aliases:
+        'account-badge-horizontal-outline',
+    ],
+    'badge-account-outline': [
+        // retro-compat v4 aliases:
+        'account-badge-outline',
+    ],
+    bed: [
+        // retro-compat v4 aliases:
+        'hotel',
+    ],
+    'book-alphabet': [
+        // retro-compat v4 aliases:
+        'dictionary',
+    ],
+    'book-cross': [
+        // retro-compat v4 aliases:
+        'bible',
+    ],
+    'book-music': [
+        // retro-compat v4 aliases:
+        'audiobook',
+    ],
+    'book-remove-multiple-outline': [
+        // retro-compat v4 aliases:
+        'square-inc',
+    ],
+    'calendar-sync': [
+        // retro-compat v4 aliases:
+        'calendar-repeat',
+    ],
+    'calendar-sync-outline': [
+        // retro-compat v4 aliases:
+        'calendar-repeat-outline',
+    ],
+    'card-account-details': [
+        // retro-compat v4 aliases:
+        'account-card-details',
+    ],
+    'card-account-details-outline': [
+        // retro-compat v4 aliases:
+        'account-card-details-outline',
+    ],
+    'card-account-details-star': [
+        // retro-compat v4 aliases:
+        'github-box',
+    ],
+    'card-account-details-star-outline': [
+        // retro-compat v4 aliases:
+        'github-face',
+    ],
+    'card-account-mail': [
+        // retro-compat v4 aliases:
+        'contact-mail',
+    ],
+    'card-account-mail-outline': [
+        // retro-compat v4 aliases:
+        'contact-mail-outline',
+    ],
+    'card-account-phone': [
+        // retro-compat v4 aliases:
+        'contact-phone',
+    ],
+    'card-account-phone-outline': [
+        // retro-compat v4 aliases:
+        'contact-phone-outline',
+    ],
+    'cellphone-cog': [
+        // retro-compat v4 aliases:
+        'cellphone-settings-variant',
+    ],
+    'circle-multiple-outline': [
+        // retro-compat v4 aliases:
+        'coins',
+    ],
+    'cloud-refresh': [
+        // retro-compat v4 aliases:
+        'tor',
+    ],
+    'code-json': [
+        // retro-compat v4 aliases:
+        'json',
+    ],
+    cog: [
+        // retro-compat v4 aliases:
+        'settings',
+    ],
+    'cog-box': [
+        // retro-compat v4 aliases:
+        'settings-box',
+    ],
+    'cog-outline': [
+        // retro-compat v4 aliases:
+        'settings-outline',
+    ],
+    'cog-transfer': [
+        // retro-compat v4 aliases:
+        'settings-transfer',
+    ],
+    'cog-transfer-outline': [
+        // retro-compat v4 aliases:
+        'settings-transfer-outline',
+    ],
+    'currency-usd-circle-outline': [
+        // retro-compat v4 aliases:
+        'coin-outline',
+    ],
+    'file-cog': [
+        // retro-compat v4 aliases:
+        'file-settings-variant',
+    ],
+    'file-cog-outline': [
+        // retro-compat v4 aliases:
+        'file-settings-variant-outline',
+    ],
+    'filmstrip-box-multiple': [
+        // retro-compat v4 aliases:
+        'library-movie',
+    ],
+    'folder-cog': [
+        // retro-compat v4 aliases:
+        'folder-settings-variant',
+    ],
+    'folder-cog-outline': [
+        // retro-compat v4 aliases:
+        'folder-settings-variant-outline',
+    ],
+    'folder-refresh-outline': [
+        // retro-compat v4 aliases:
+        'tumblr-reblog',
+    ],
+    'form-textarea': [
+        // retro-compat v4 aliases:
+        'textarea',
+    ],
+    'form-textbox': [
+        // retro-compat v4 aliases:
+        'textbox',
+    ],
+    'form-textbox-password': [
+        // retro-compat v4 aliases:
+        'textbox-password',
+    ],
+    github: [
+        // retro-compat v4 aliases:
+        'github-circle',
+    ],
+    'ice-cream-off': [
+        // retro-compat v4 aliases:
+        'nfc-off',
+    ],
+    'image-multiple-outline': [
+        // retro-compat v4 aliases:
+        'image-filter',
+    ],
+    'language-markdown': [
+        // retro-compat v4 aliases:
+        'markdown',
+    ],
+    'language-markdown-outline': [
+        // retro-compat v4 aliases:
+        'markdown-outline',
+    ],
+    'language-ruby': [
+        // retro-compat v4 aliases:
+        'ruby',
+    ],
+    'language-xaml': [
+        // retro-compat v4 aliases:
+        'xaml',
+    ],
+    'message-cog': [
+        // retro-compat v4 aliases:
+        'message-settings-variant',
+    ],
+    'microsoft-azure': [
+        // retro-compat v4 aliases:
+        'azure',
+    ],
+    'microsoft-azure-devops': [
+        // retro-compat v4 aliases:
+        'azure-devops',
+    ],
+    'microsoft-bing': [
+        // retro-compat v4 aliases:
+        'bing',
+    ],
+    'microsoft-dynamics-365': [
+        // retro-compat v4 aliases:
+        'microsoft-dynamics',
+    ],
+    'microsoft-edge': [
+        // retro-compat v4 aliases:
+        'edge',
+    ],
+    'microsoft-internet-explorer': [
+        // retro-compat v4 aliases:
+        'internet-explorer',
+    ],
+    'microsoft-office': [
+        // retro-compat v4 aliases:
+        'office',
+    ],
+    'microsoft-onedrive': [
+        // retro-compat v4 aliases:
+        'onedrive',
+    ],
+    'microsoft-onenote': [
+        // retro-compat v4 aliases:
+        'onenote',
+    ],
+    'microsoft-outlook': [
+        // retro-compat v4 aliases:
+        'outlook',
+    ],
+    'microsoft-visual-studio': [
+        // retro-compat v4 aliases:
+        'visual-studio',
+    ],
+    'microsoft-visual-studio-code': [
+        // retro-compat v4 aliases:
+        'visual-studio-code',
+    ],
+    'microsoft-windows': [
+        // retro-compat v4 aliases:
+        'windows',
+    ],
+    'microsoft-windows-classic': [
+        // retro-compat v4 aliases:
+        'windows-classic',
+    ],
+    'microsoft-xbox': [
+        // retro-compat v4 aliases:
+        'xbox',
+    ],
+    'microsoft-xbox-controller': [
+        // retro-compat v4 aliases:
+        'xbox-controller',
+    ],
+    'microsoft-xbox-controller-battery-alert': [
+        // retro-compat v4 aliases:
+        'xbox-controller-battery-alert',
+    ],
+    'microsoft-xbox-controller-battery-charging': [
+        // retro-compat v4 aliases:
+        'xbox-controller-battery-charging',
+    ],
+    'microsoft-xbox-controller-battery-empty': [
+        // retro-compat v4 aliases:
+        'xbox-controller-battery-empty',
+    ],
+    'microsoft-xbox-controller-battery-full': [
+        // retro-compat v4 aliases:
+        'xbox-controller-battery-full',
+    ],
+    'microsoft-xbox-controller-battery-low': [
+        // retro-compat v4 aliases:
+        'xbox-controller-battery-low',
+    ],
+    'microsoft-xbox-controller-battery-medium': [
+        // retro-compat v4 aliases:
+        'xbox-controller-battery-medium',
+    ],
+    'microsoft-xbox-controller-battery-unknown': [
+        // retro-compat v4 aliases:
+        'xbox-controller-battery-unknown',
+    ],
+    'microsoft-xbox-controller-menu': [
+        // retro-compat v4 aliases:
+        'xbox-controller-menu',
+    ],
+    'microsoft-xbox-controller-off': [
+        // retro-compat v4 aliases:
+        'xbox-controller-off',
+    ],
+    'microsoft-xbox-controller-view': [
+        // retro-compat v4 aliases:
+        'xbox-controller-view',
+    ],
+    'microsoft-yammer': [
+        // retro-compat v4 aliases:
+        'yammer',
+    ],
+    'molecule-co2': [
+        // retro-compat v4 aliases:
+        'periodic-table-co2',
+    ],
+    'music-box-multiple': [
+        // retro-compat v4 aliases:
+        'library-music',
+    ],
+    'music-box-multiple-outline': [
+        // retro-compat v4 aliases:
+        'library-music-outline',
+    ],
+    'nintendo-wii': [
+        // retro-compat v4 aliases:
+        'wii',
+    ],
+    'nintendo-wiiu': [
+        // retro-compat v4 aliases:
+        'wiiu',
+    ],
+    facebook: [
+        // retro-compat v4 aliases:
+        'facebook-box',
+    ],
+    'google-plus': [
+        // retro-compat v4 aliases:
+        'google-plus-box',
+    ],
+    npm: [
+        // retro-compat v4 aliases:
+        'npm-variant',
+        'npm-variant-outline',
+    ],
+    'x-twitter': [
+        'twitter',
+        // retro-compat v4 aliases:
+        'twitter-box',
+        'twitter-circle',
+    ],
+    'play-box-multiple': [
+        // retro-compat v4 aliases:
+        'library-video',
+    ],
+    'roller-skate-off': [
+        // retro-compat v4 aliases:
+        'cisco-webex',
+    ],
+    'router-network': [
+        // retro-compat v4 aliases:
+        'network-router',
+    ],
+    'sail-boat': [
+        // retro-compat v4 aliases:
+        'sailing',
+    ],
+    'sony-playstation': [
+        // retro-compat v4 aliases:
+        'playstation',
+    ],
+    'sort-alphabetical-variant': [
+        // retro-compat v4 aliases:
+        'sort-alphabetical',
+    ],
+    'sort-numeric-variant': [
+        // retro-compat v4 aliases:
+        'sort-numeric',
+    ],
+    linkedin: [
+        // retro-compat v4 aliases:
+        'linkedin-box',
+    ],
+    'text-box': [
+        // retro-compat v4 aliases:
+        'file-document-box',
+    ],
+    'text-box-check': [
+        // retro-compat v4 aliases:
+        'file-document-box-check',
+    ],
+    'text-box-check-outline': [
+        // retro-compat v4 aliases:
+        'file-document-box-check-outline',
+    ],
+    'text-box-minus': [
+        // retro-compat v4 aliases:
+        'file-document-box-minus',
+    ],
+    'text-box-minus-outline': [
+        // retro-compat v4 aliases:
+        'file-document-box-minus-outline',
+    ],
+    'text-box-multiple': [
+        // retro-compat v4 aliases:
+        'library-books',
+        'file-document-box-multiple',
+    ],
+    'text-box-multiple-outline': [
+        // retro-compat v4 aliases:
+        'file-document-box-multiple-outline',
+    ],
+    'text-box-outline': [
+        // retro-compat v4 aliases:
+        'file-document-box-outline',
+    ],
+    'text-box-plus': [
+        // retro-compat v4 aliases:
+        'file-document-box-plus',
+    ],
+    'text-box-plus-outline': [
+        // retro-compat v4 aliases:
+        'file-document-box-plus-outline',
+    ],
+    'text-box-remove': [
+        // retro-compat v4 aliases:
+        'file-document-box-remove',
+    ],
+    'text-box-remove-outline': [
+        // retro-compat v4 aliases:
+        'file-document-box-remove-outline',
+    ],
+    'text-box-search': [
+        // retro-compat v4 aliases:
+        'file-document-box-search',
+    ],
+    'text-box-search-outline': [
+        // retro-compat v4 aliases:
+        'file-document-box-search-outline',
+    ],
+    'tow-truck': [
+        // retro-compat v4 aliases:
+        'towing',
+    ],
+    'video-box': [
+        // retro-compat v4 aliases:
+        'camcorder-box',
+    ],
+    'video-box-off': [
+        // retro-compat v4 aliases:
+        'camcorder-box-off',
+    ],
+    'wall-sconce-round-variant': [
+        // retro-compat v4 aliases:
+        'wall-sconce-variant',
+    ],
+    'y-combinator': [
+        // retro-compat v4 aliases:
+        'hackernews',
+    ],
+    'youtube-studio': [
+        // retro-compat v4 aliases:
+        'youtube-creator-studio',
+    ],
+    'format-title': [
+        // retro-compat v4 aliases:
+        'instapaper',
+    ],
+    at: [
+        // retro-compat v4 aliases:
+        'mail-ru',
+    ],
+};